requirements.txt
aniso8601==7.0.0
backcall==0.2.0
click==7.1.2
click-log==0.3.2
dataclasses==0.6
dataclasses-json==0.5.2
decorator==4.4.2
fb-sapp==0.2.8
Flask==1.1.2
Flask-GraphQL==2.0.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
libcst==0.3.16
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
munch==2.5.0
mypy-extensions==0.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.14
psutil==5.8.0
ptyprocess==0.7.0
Pygments==2.7.4
pyre-check==0.0.59
pyre-extensions==0.0.20
python-dotenv==0.15.0
pywatchman==1.4.1
PyYAML==5.4.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.23
stringcase==1.2.0
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ujson==4.0.2
wcwidth==0.2.5
Werkzeug==1.0.1
xxhash==2.0.0
